+++ src/gov/nasa/gsfc/spdf/cdfj/DataTypes.java
@@ -111,7 +111,7 @@ public final class DataTypes {
     }
     public static String getString(ByteBuffer buf, Integer nc)  {
         ByteBuffer slice = buf.slice();
-        byte [] ba = new byte[nc.intValue()];
+        byte [] ba = new byte[nc];
         int i = 0;
         for (; i < ba.length; i++) {
             ba[i] = slice.get();
