+++ src/gov/nasa/gsfc/spdf/cdfj/DoubleVarContainer.java
@@ -190,14 +190,14 @@ public final class DoubleVarContainer extends BaseVarContainer implements
             _buf.get(_a0);
             return (singlePoint)?new Double(_a0[0]):_a0;
         case 1:
-            int n = (((Integer)var.getElementCount().elementAt(0))).intValue();
+            int n = (((Integer)var.getElementCount().elementAt(0)));
             records = words/n;
             double[][] _a1 = new double[records][n];
             for (int r = 0; r < records; r++) _buf.get(_a1[r]);
             return (singlePoint)?_a1[0]:_a1;
         case 2:
-            int n0 = (((Integer)var.getElementCount().elementAt(0))).intValue();
-            int n1 = (((Integer)var.getElementCount().elementAt(1))).intValue();
+            int n0 = (((Integer)var.getElementCount().elementAt(0)));
+            int n1 = (((Integer)var.getElementCount().elementAt(1)));
             records = words/(n0*n1);
             double[][][] _a2 = new double[records][n0][n1];
             if (var.rowMajority()) {
@@ -215,9 +215,9 @@ public final class DoubleVarContainer extends BaseVarContainer implements
             }
             return (singlePoint)?_a2[0]:_a2;
         case 3:
-            n0 = (((Integer)var.getElementCount().elementAt(0))).intValue();
-            n1 = (((Integer)var.getElementCount().elementAt(1))).intValue();
-            int n2 = (((Integer)var.getElementCount().elementAt(2))).intValue();
+            n0 = (((Integer)var.getElementCount().elementAt(0)));
+            n1 = (((Integer)var.getElementCount().elementAt(1)));
+            int n2 = (((Integer)var.getElementCount().elementAt(2)));
             records = words/(n0*n1*n2);
             double[][][][] _a3 = new double[records][n0][n1][n2];
             if (var.rowMajority()) {
@@ -241,10 +241,10 @@ public final class DoubleVarContainer extends BaseVarContainer implements
             }
             return (singlePoint)?_a3[0]:_a3;
         case 4:
-            n0 = (((Integer)var.getElementCount().elementAt(0))).intValue();
-            n1 = (((Integer)var.getElementCount().elementAt(1))).intValue();
-            n2 = (((Integer)var.getElementCount().elementAt(2))).intValue();
-            int n3 = (((Integer)var.getElementCount().elementAt(3))).intValue();
+            n0 = (((Integer)var.getElementCount().elementAt(0)));
+            n1 = (((Integer)var.getElementCount().elementAt(1)));
+            n2 = (((Integer)var.getElementCount().elementAt(2)));
+            int n3 = (((Integer)var.getElementCount().elementAt(3)));
             records = words/(n0*n1*n2*n3);
             double[][][][][] _a4 = new double[records][n0][n1][n2][n3];
             if (var.rowMajority()) {
@@ -285,7 +285,7 @@ public final class DoubleVarContainer extends BaseVarContainer implements
         int words = (buf.remaining())/8;
         DoubleBuffer _buf = buf.asDoubleBuffer();
         int records = -1;
-        int n = (((Integer)var.getElementCount().elementAt(0))).intValue();
+        int n = (((Integer)var.getElementCount().elementAt(0)));
         records = words/n;
         if (elements.length == 1) {
             int element = elements[0];
@@ -313,11 +313,11 @@ public final class DoubleVarContainer extends BaseVarContainer implements
     public Object asArrayElement(int index0, int index1) throws Throwable {
         int rank = var.getEffectiveRank();
         if (rank != 2) throw new Throwable("Rank other than 2 not supported.");
-        int n0 = (((Integer)var.getElementCount().elementAt(0))).intValue();
+        int n0 = (((Integer)var.getElementCount().elementAt(0)));
         if ((index0 < 0) || (index0 >= n0)) {
             throw new Throwable("Invalid first index " + index0);
         }
-        int n1 = (((Integer)var.getElementCount().elementAt(1))).intValue();
+        int n1 = (((Integer)var.getElementCount().elementAt(1)));
         if ((index1 < 0) || (index1 >= n1)) {
             throw new Throwable("Invalid second index " + index1);
         }
